+++ src/Products/PluginIndexes/util.py
@@ -11,7 +11,12 @@
 #
 ##############################################################################
 
+from datetime import datetime
+
+from DateTime.DateTime import DateTime
 import six
+
+MAX32 = int(2 ** 31 - 1)
 
 
 def safe_callable(ob):
@@ -20,3 +25,28 @@ def safe_callable(ob):
         return hasattr(ob, '__call__') or isinstance(ob, six.class_types)
     else:
         return callable(ob)
+
+
+def datetime_to_minutes(value, precision=1,
+                        max_value=MAX32, min_value=-MAX32):
+    if value is None:
+        return value
+
+    if isinstance(value, (str, datetime)):
+        value = DateTime(value)
+
+    if isinstance(value, DateTime):
+        value = value.millis() / 1000 / 60  # flatten to minutes
+
+    # flatten to precision
+    if precision > 1:
+        value = value - (value % precision)
+
+    value = int(value)
+
+    if value > max_value or value < min_value:
+        # value must be integer fitting in the range (default 32bit)
+        raise OverflowError(
+            '%s is not within the range of dates allowed.' % value)
+
+    return value